Align FlashAttention with Week 2 BF16
Keep the model-facing GPU path and output in BF16 while retaining FP32 for scale, masks, and accumulation, and make BF16 the default D=128 GPU test contract.

+++ src/tiny_llm_ref/attention.py
@@ -121,7 +121,7 @@ def flash_attention(
     mask: mx.array | str | None = None,
 ) -> mx.array:
     factor = mx.rsqrt(query.shape[-1]) if scale is None else mx.array(scale)
-    factor = factor.astype(query.dtype)
+    factor = factor.astype(mx.float32)
 
     *B, H_q, L, E = query.shape
     _, H, S, _ = key.shape
